Moving UpdateExistingResourceData to the environment specific level
