+++ pkg/corerp/frontend/controller/environments/createorupdateenvironment.go
@@ -60,7 +60,7 @@ func (e *CreateOrUpdateEnvironment) Run(ctx context.Context, req *http.Request) 
 		return rest.NewPreconditionFailedResponse(err.Error()), nil
 	}
 
-	ctrl.UpdateExistingResourceData(ctx, existingResource, newResource)
+	UpdateExistingResourceData(ctx, existingResource, newResource)
 
 	nr, err := e.SaveResource(ctx, serviceCtx.ResourceID.ID, newResource, etag)
 	if err != nil {
@@ -105,4 +105,14 @@ func (e *CreateOrUpdateEnvironment) Validate(ctx context.Context, req *http.Requ
 	dm.Properties.ProvisioningState = datamodel.ProvisioningStateSucceeded
 
 	return dm, err
+}
+
+// UpdateExistingResourceData updates the environment resource before it is saved to the DB.
+func UpdateExistingResourceData(ctx context.Context, er *datamodel.Environment, nr *datamodel.Environment) {
+	sc := servicecontext.ARMRequestContextFromContext(ctx)
+	nr.SystemData = ctrl.UpdateSystemData(er.SystemData, *sc.SystemData())
+	if er.CreatedAPIVersion != "" {
+		nr.CreatedAPIVersion = er.CreatedAPIVersion
+	}
+	nr.TenantID = sc.HomeTenantID
 }
